WannaBeYours/requirements.txt
torch
torchvision
Pillow
requests
gitpython
matplotlib
